[Refactoring ⚙️] 레이아웃 좌/우 컴포넌트 파일 분리
