+++ src/components/Navbar/Navbar.Styles.ts
@@ -10,30 +10,8 @@ export const NavbarLayout = styled.section`
   border-bottom: 0.1rem solid #000;
 `
 
-export const NavbarLeftList = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-`
-
 export const NavbarButton = styled.button`
   svg {
     font-size: 3rem;
   }
-`
-export const NavbarLogo = styled.img`
-  width: 10rem;
-  height: 10rem;
-  border-radius: 50%;
-`
-
-export const NavbarRightList = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 2rem;
-`
-export const NavbarProfile = styled.img`
-  width: 6rem;
-  height: 6rem;
-  border-radius: 50%;
 `
