+++ src/components/Navbar/components/NavbarLeftList/NavbarLeftList.Styles.ts
@@ -1,0 +1,13 @@
+import styled from "styled-components"
+
+export const NavbarLeftListLayout = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+`
+
+export const NavbarLogo = styled.img`
+  width: 10rem;
+  height: 10rem;
+  border-radius: 50%;
+`
