+++ src/components/Navbar/components/NavbarRightList/NavbarRightList.Styles.ts
@@ -1,0 +1,13 @@
+import styled from "styled-components"
+
+export const NavbarRightListlayout = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 2rem;
+`
+
+export const NavbarProfile = styled.img`
+  width: 6rem;
+  height: 6rem;
+  border-radius: 50%;
+`
